Refactor: Calendar 변수명 ActivityCalendar와 StatusCalendar로 분리
